Rename __subTypes field to __fragments
The __subTypes field didn't actually map to the different sub-types in
the abstract GraphQL interface, but actually to each fragment used in
that selection.

I think we should rename to __fragments instead of merging the same
sub-type into one type because the person writing the GraphQL query
can visibly inspect which fragment they're trying to query, instead of
needing to figure out which order their types are in.